[WIP] Fixed issue to keep focus inside mobile menu
focus should be trapped inside the modal. Once you tab past the last menu item, the focus should go on the close button. Once you shift+tab on the first element focus should go on the close button (and if you shift+tab on this, it should focus the last element in the menu).